chore: add events hook

+++ Hardhat/contracts/Forum.sol
@@ -5,48 +5,48 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 contract Forum {
     struct Question {
-        uint questionId;
+        uint256 questionId;
         string message;
         address creatorAddress;
-        uint timestamp;
+        uint256 timestamp;
     }
 
     struct Answer {
-        uint answerId;
-        uint questionId;
+        uint256 answerId;
+        uint256 questionId;
         string message;
         address creatorAddress;
-        uint timestamp;
-        uint upvotes;
+        uint256 timestamp;
+        uint256 upvotes;
     }
 
     Question[] public questions;
     Answer[] public answers;
-    mapping(uint => uint[]) public answerPerQuestion;
-    mapping(uint => mapping(address => bool)) public upvoters;
-    mapping(address => uint) public userUpvoteCount;
+    mapping(uint256 => uint256[]) public answerPerQuestion;
+    mapping(uint256 => mapping(address => bool)) public upvoters;
+    mapping(address => uint256) public userUpvoteCount;
 
     event QuestionAdded(Question question);
     event AnswerAdded(Answer answer);
     event AnswerUpvoted(Answer answer);
 
     IERC20 public immutable Goflow;
-    uint constant decimal = 18;
+    uint256 constant decimal = 18;
 
-    uint amountToPay = 1 * 10 ** decimal;
-    uint amountToPayParticipate = 10 * 10 ** decimal;
+    uint256 amountToPay = 1 * 10**decimal;
+    uint256 amountToPayParticipate = 10 * 10**decimal;
 
     constructor(address _tokenAddress) {
         Goflow = IERC20(_tokenAddress);
     }
 
-    modifier answerExists(uint _answerId) {
+    modifier answerExists(uint256 _answerId) {
         require(answers.length >= _answerId, "Answer does not exist");
         _;
     }
 
     function postQuestion(string calldata _message) external {
-        uint questionCounter = questions.length;
+        uint256 questionCounter = questions.length;
         Question memory question = Question({
             questionId: questionCounter,
             message: _message,
@@ -57,8 +57,8 @@ contract Forum {
         emit QuestionAdded(question);
     }
 
-    function postAnswer(uint questionId, string calldata _message) external {
-        uint answerCounter = answers.length;
+    function postAnswer(uint256 questionId, string calldata _message) external {
+        uint256 answerCounter = answers.length;
         Answer memory answer = Answer({
             answerId: answerCounter,
             questionId: questionId,
@@ -72,7 +72,7 @@ contract Forum {
         emit AnswerAdded(answer);
     }
 
-    function upvoteAnswer(uint _answerId) external answerExists(_answerId) {
+    function upvoteAnswer(uint256 _answerId) external answerExists(_answerId) {
         Answer storage currentAnswer = answers[_answerId];
 
         require(
@@ -113,9 +113,12 @@ contract Forum {
         emit AnswerUpvoted(currentAnswer);
     }
 
-    function getUpvotes(
-        uint _answerId
-    ) public view answerExists(_answerId) returns (uint) {
+    function getUpvotes(uint256 _answerId)
+        public
+        view
+        answerExists(_answerId)
+        returns (uint256)
+    {
         return answers[_answerId].upvotes;
     }
 
@@ -123,9 +126,11 @@ contract Forum {
         return questions;
     }
 
-    function getAnswersPerQuestion(
-        uint _questionId
-    ) public view returns (uint[] memory) {
+    function getAnswersPerQuestion(uint256 _questionId)
+        public
+        view
+        returns (uint256[] memory)
+    {
         return answerPerQuestion[_questionId];
     }
 }
